+++ ensmovielibrary/src/styles/App.css
@@ -23,3 +23,8 @@ h1 {
 	width: 150px;
 	height: 200px;
 }
+
+.arrowImage {
+	width: 50px;
+	height: 50px;
+}
