+++ ensmovielibrary/src/styles/NavBar.css
@@ -70,8 +70,3 @@
 	height: 50px;
 	margin-left: 10px;
 }
-
-.arrowImage {
-	width: 50px;
-	height: 50px;
-}
